fix(airflow): fix schema
